refactor: isolate automation security chain from management security
