+++ gravitee-am-management-api/gravitee-am-management-api-standalone/gravitee-am-management-api-standalone-container/pom.xml
@@ -90,12 +90,6 @@
         </dependency>
 
         <dependency>
-            <groupId>io.gravitee.am.management.automation</groupId>
-            <artifactId>gravitee-am-management-api-automation-security</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
             <groupId>io.gravitee.am</groupId>
             <artifactId>gravitee-am-monitoring</artifactId>
             <version>${project.version}</version>
